+++ examples/static-site-generator/playground-export-smoke.mjs
@@ -440,6 +440,22 @@ $schema_publisher_id = wp_insert_post(array(
 	'post_content' => '<p>Schema publisher export target.</p>',
 ));
 
+$schema_contributor_id = wp_insert_post(array(
+	'post_type' => 'page',
+	'post_status' => 'publish',
+	'post_title' => 'Schema Contributor',
+	'post_name' => 'schema-contributor',
+	'post_content' => '<p>Schema contributor export target.</p>',
+));
+
+$schema_reviewer_id = wp_insert_post(array(
+	'post_type' => 'page',
+	'post_status' => 'publish',
+	'post_title' => 'Schema Reviewer',
+	'post_name' => 'schema-reviewer',
+	'post_content' => '<p>Schema reviewer export target.</p>',
+));
+
 $amp_id = wp_insert_post(array(
 	'post_type' => 'page',
 	'post_status' => 'publish',
@@ -497,6 +513,8 @@ $schema_source_url = get_permalink($schema_source_id);
 $schema_policy_url = get_permalink($schema_policy_id);
 $schema_author_url = get_permalink($schema_author_id);
 $schema_publisher_url = get_permalink($schema_publisher_id);
+$schema_contributor_url = get_permalink($schema_contributor_id);
+$schema_reviewer_url = get_permalink($schema_reviewer_id);
 $amp_url = get_permalink($amp_id);
 $preloaded_document_url = get_permalink($preloaded_document_id);
 $protocol_child_url = preg_replace('/^https?:/', '', $child_url);
@@ -564,6 +582,8 @@ $static_content = '<p id="section">Static smoke page.</p>'
 	. '<link itemprop="acquireLicensePage" href="' . esc_url($microdata_license_url) . '">'
 	. '<link itemprop="author" href="' . esc_url($schema_author_url) . '">'
 	. '<meta itemprop="publisher" content="' . esc_url($schema_publisher_url) . '">'
+	. '<link itemprop="contributor" href="' . esc_url($schema_contributor_url) . '">'
+	. '<meta itemprop="reviewedBy" content="' . esc_url($schema_reviewer_url) . '">'
 	. '<link itemprop="contentUrl" href="' . esc_url($asset_url . '?schema-link=1') . '">'
 	. '<article itemscope itemid="' . esc_url($microdata_item_url) . '" itemtype="'
 	. esc_url($microdata_type_url) . ' https://schema.org/Article '
@@ -710,6 +730,8 @@ $scoped_schema_source_url = get_permalink($schema_source_id);
 $scoped_schema_policy_url = get_permalink($schema_policy_id);
 $scoped_schema_author_url = get_permalink($schema_author_id);
 $scoped_schema_publisher_url = get_permalink($schema_publisher_id);
+$scoped_schema_contributor_url = get_permalink($schema_contributor_id);
+$scoped_schema_reviewer_url = get_permalink($schema_reviewer_id);
 $scoped_amp_url = get_permalink($amp_id);
 $scoped_preloaded_document_url = get_permalink($preloaded_document_id);
 $scoped_protocol_child_url = preg_replace('/^https?:/', '', $scoped_child_url);
@@ -786,6 +808,8 @@ $scoped_static_content = '<p id="section">Static smoke page.</p>'
 	. '<link itemprop="acquireLicensePage" href="' . esc_url($scoped_microdata_license_url) . '">'
 	. '<link itemprop="author" href="' . esc_url($scoped_schema_author_url) . '">'
 	. '<meta itemprop="publisher" content="' . esc_url($scoped_schema_publisher_url) . '">'
+	. '<link itemprop="contributor" href="' . esc_url($scoped_schema_contributor_url) . '">'
+	. '<meta itemprop="reviewedBy" content="' . esc_url($scoped_schema_reviewer_url) . '">'
 	. '<link itemprop="contentUrl" href="' . esc_url($scoped_asset_url . '?schema-link=1') . '">'
 	. '<article itemscope itemid="' . esc_url($scoped_microdata_item_url) . '" itemtype="'
 	. esc_url($scoped_microdata_type_url) . ' https://schema.org/Article '
@@ -975,6 +999,8 @@ async function verifyExport() {
 	assertFile('publishing-principles/index.html');
 	assertFile('schema-author/index.html');
 	assertFile('schema-publisher/index.html');
+	assertFile('schema-contributor/index.html');
+	assertFile('schema-reviewer/index.html');
 	assertFile('amp-companion/index.html');
 	assertFile('preloaded-document/index.html');
 	assertFile('parent-page/index.html');
@@ -1037,6 +1063,8 @@ async function verifyExport() {
 		'../publishing-principles/index.html',
 		'../schema-author/index.html',
 		'../schema-publisher/index.html',
+		'../schema-contributor/index.html',
+		'../schema-reviewer/index.html',
 		'../amp-companion/index.html',
 		'../preloaded-document/index.html',
 		'relative-child/index.html',
@@ -1201,6 +1229,16 @@ async function verifyExport() {
 		staticPage,
 		'<meta itemprop="publisher" content="../schema-publisher/index.html">',
 		'static-page/index.html rewrites schema.org publisher metadata'
+	);
+	assertIncludes(
+		staticPage,
+		'<link itemprop="contributor" href="../schema-contributor/index.html">',
+		'static-page/index.html rewrites schema.org contributor links'
+	);
+	assertIncludes(
+		staticPage,
+		'<meta itemprop="reviewedBy" content="../schema-reviewer/index.html">',
+		'static-page/index.html rewrites schema.org reviewedBy metadata'
 	);
 	assertIncludes(
 		staticPage,
@@ -1492,6 +1530,8 @@ async function verifyScopedExport() {
 	assertFile('publishing-principles/index.html');
 	assertFile('amp-companion/index.html');
 	assertFile('preloaded-document/index.html');
+	assertFile('schema-contributor/index.html');
+	assertFile('schema-reviewer/index.html');
 	assertFile('parent-page/child-page/index.html');
 	assertFile('wp-content/uploads/ssgwp-smoke-asset.txt');
 	assertFile('wp-content/uploads/ssgwp-smoke-captions.vtt');
@@ -1562,6 +1602,8 @@ async function verifyScopedExport() {
 		'../schema-part/index.html',
 		'../schema-source/index.html',
 		'../publishing-principles/index.html',
+		'../schema-contributor/index.html',
+		'../schema-reviewer/index.html',
 		'../amp-companion/index.html',
 		'../preloaded-document/index.html',
 		'relative-child/index.html',
@@ -1720,6 +1762,16 @@ async function verifyScopedExport() {
 		staticPage,
 		'<link itemprop="publishingPrinciples" href="../publishing-principles/index.html">',
 		'scoped static-page/index.html rewrites schema.org publishingPrinciples links'
+	);
+	assertIncludes(
+		staticPage,
+		'<link itemprop="contributor" href="../schema-contributor/index.html">',
+		'scoped static-page/index.html rewrites schema.org contributor links'
+	);
+	assertIncludes(
+		staticPage,
+		'<meta itemprop="reviewedBy" content="../schema-reviewer/index.html">',
+		'scoped static-page/index.html rewrites schema.org reviewedBy metadata'
 	);
 	assertIncludes(
 		staticPage,